@@ -1,11 +1,11 @@
 module github.com/zerodha/gokiteconnect/v4
 
-go 1.14
+go 1.16
 
 require (
-	github.com/gocarina/gocsv v0.0.0-20180809181117-b8c38cb1ba36
-	github.com/google/go-querystring v1.0.0
-	github.com/gorilla/websocket v1.4.2
-	github.com/stretchr/testify v1.7.0
+	github.com/gocarina/gocsv v0.0.0-20220927221512-ad3251f9fa25
+	github.com/google/go-querystring v1.1.0
+	github.com/gorilla/websocket v1.5.0
+	github.com/stretchr/testify v1.8.0
 	gopkg.in/jarcoal/httpmock.v1 v1.0.0-20180719183105-8007e27cdb32
 )
